refactor(json-schema): extract conversion modules
