+++ lib/base.js
@@ -10,47 +10,14 @@ const Extend = require('./extend');
 const Manifest = require('./manifest');
 const Messages = require('./messages');
 const Modify = require('./modify');
+const JsonSchema = require('./json-schema/base');
 const Ref = require('./ref');
 const Trace = require('./trace');
 const Validator = require('./validator');
 const Values = require('./values');
 
 
-const internals = {
-    standardTypes: new Set(['string', 'number', 'integer', 'boolean', 'object', 'array', 'null']),
-    jsonSchemaTarget: 'draft-2020-12',
-    primitiveTypes: new Set(['string', 'number', 'boolean']),
-    // Keywords copied through from user-provided meta() values.
-    metaPassthroughKeywords: new Set([
-        '$comment',
-        'contentEncoding',
-        'contentMediaType',
-        'contentSchema',
-        'deprecated',
-        'examples',
-        'format',
-        'readOnly',
-        'title',
-        'writeOnly'
-    ]),
-    // Keywords preserved when exclusive valids have to fall back to enum-only
-    // output. This intentionally excludes validating keywords such as format.
-    onlyFallbackAnnotationKeywords: new Set([
-        '$comment',
-        '$defs',
-        'contentEncoding',
-        'contentMediaType',
-        'contentSchema',
-        'default',
-        'deprecated',
-        'description',
-        'examples',
-        'readOnly',
-        'title',
-        'writeOnly'
-    ]),
-    nullSchema: () => ({ type: 'null' })
-};
+const internals = {};
 
 
 internals.Base = class {
@@ -98,248 +65,7 @@ internals.Base = class {
 
     $_jsonSchema(mode, options = {}) {
 
-        if (options.target !== undefined &&
-            options.target !== internals.jsonSchemaTarget) {
-
-            throw new Error(`Unsupported JSON Schema target: ${options.target}`);
-        }
-
-        const rootCall = !options.$defs;
-        const defs = options.$defs ?? {};
-
-        // Merge parent prefs with this schema's prefs (explicit takes precedence)
-
-        const prefs = this._preferences
-            ? Common.preferences(options.prefs, this._preferences)
-            : options.prefs;
-        const presence = this._flags.presence || prefs?.presence;
-
-        if (presence === 'forbidden' && !options.ignorePresence) {
-            return false;
-        }
-
-        let schema = {};
-
-        const jsonSchemaType = internals.jsonSchemaType(this);
-        const isTypeAny = this.type === 'any';
-        const isOnly = this._flags.only;
-
-        const rawValids = this._valids && Array.from(this._valids._values);
-        const valids = rawValids && internals.jsonSchemaValues(this, rawValids);
-        const onlyValues = valids && valids.filter((v) => typeof v !== 'symbol');
-        const nonNullValids = valids && valids.filter((v) => v !== null);
-        const rawAllowedValues = rawValids && rawValids.filter((v) => typeof v !== 'symbol' && v !== null);
-        let typesOverlap = true;
-
-        // If 'only' is set, check if the allowed values' types overlap with the schema type
-
-        if (rawValids && isOnly && !isTypeAny) {
-            const comparableValues = rawValids.filter((v) => typeof v !== 'symbol' && v !== null);
-            if (!comparableValues.length) {
-                typesOverlap = true;
-            }
-            else {
-                const types = new Set(comparableValues.map((v) => typeof v));
-                const type = jsonSchemaType || this.type;
-                typesOverlap = types.has(type) || (type === 'date' && types.has('object'));
-            }
-        }
-
-        // Set the JSON Schema 'type' if it's a standard type and there's an overlap
-
-        if (!isTypeAny && typesOverlap && jsonSchemaType) {
-            schema.type = jsonSchemaType;
-        }
-
-        if (this._flags.description) {
-            schema.description = this._flags.description;
-        }
-
-        if (this._flags.default !== undefined && typeof this._flags.default !== 'function') {
-            const defaultValue = internals.jsonSchemaDefaultValue(this, this._flags.default);
-            if (defaultValue !== undefined) {
-                schema.default = defaultValue;
-            }
-        }
-
-        // Apply type-specific JSON Schema conversion
-
-        const subOptions = { ...options, $defs: defs, prefs };
-        if (this._definition.jsonSchema && typesOverlap) {
-            schema = this._definition.jsonSchema(this, schema, mode, subOptions);
-        }
-
-        // Apply rule-specific JSON Schema conversions
-
-        for (const rule of this._rules) {
-            const definition = this._definition.rules[rule.name];
-            if (definition.jsonSchema && typesOverlap && !rule._resolve.length) {
-                schema = definition.jsonSchema(rule, schema, isOnly, mode, subOptions);
-            }
-        }
-
-        // Handle shared schemas
-
-        if (this.$_terms.shared) {
-            for (const shared of this.$_terms.shared) {
-                defs[shared._flags.id] = shared.$_jsonSchema(mode, subOptions);
-            }
-        }
-
-        internals.applyExamples(this, schema, this.$_terms.examples);
-        internals.applyMetas(this, schema, this.$_terms.metas);
-
-        if (rootCall && Object.keys(defs).length) {
-            schema.$defs = defs;
-        }
-
-        // Handle allowed values (valids)
-
-        if (this._valids) {
-
-            const values = isOnly ? onlyValues : nonNullValids.filter((v) => typeof v !== 'symbol');
-            if (values.length) {
-                if (this._flags.only) {
-                    if (!(values.length === 1 && values[0] === null)) {
-                        if (typeof schema !== 'boolean') {
-                            schema.enum = values;
-
-                            const types = internals.onlyTypes(values);
-
-                            if (types) {
-                                schema.type = types.length === 1 ? types[0] : types;
-                            }
-                        }
-                    }
-                }
-                else {
-                    // If values are allowed but not exclusive, add them via 'anyOf' when
-                    // the base schema would otherwise reject them.
-
-                    const extras = [];
-                    if (!isTypeAny) {
-                        const base = internals.onlyBaseClone(this);
-                        const schemaTypes = new Set(internals.schemaTypes(schema) || []);
-                        schemaTypes.add(jsonSchemaType || this.type);
-
-                        for (const rawValue of rawAllowedValues) {
-                            const representations = internals.jsonSchemaRepresentations(this, rawValue);
-                            const candidates = [];
-
-                            for (const value of representations) {
-                                const valueType = internals.jsonSchemaValueType(value);
-                                if (valueType === null ||
-                                    !schemaTypes.has(valueType)) {
-
-                                    extras.push(value);
-                                    continue;
-                                }
-
-                                candidates.push(value);
-                            }
-
-                            if (candidates.length &&
-                                base.validate(rawValue, prefs).error) {
-
-                                extras.push(...candidates);
-                            }
-                        }
-                    }
-
-                    const uniqueExtras = internals.uniqueJsonSchemaValues(extras);
-                    if (uniqueExtras.length) {
-                        if (!schema.anyOf) {
-                            schema = {
-                                anyOf: [schema]
-                            };
-                        }
-
-                        schema.anyOf.push({ enum: uniqueExtras });
-                    }
-                }
-            }
-        }
-
-        // Handle disallowed values (invalids)
-
-        if (this._invalids) {
-            const invalids = internals.jsonSchemaValues(this, Array.from(this._invalids._values)
-                .filter((v) => typeof v !== 'symbol'))
-                .filter((v) => v !== null || internals.schemaCanMatchNull(schema));
-
-            if (invalids.length) {
-                schema = internals.appendCompositeKeyword(schema, 'not', { enum: invalids });
-            }
-        }
-
-        // Handle 'null' if it's an allowed value
-
-        if (!isOnly && this._valids && this._valids.has(null) && !isTypeAny) {
-            if (schema.type) {
-                schema.type = internals.appendType(schema.type, 'null');
-            }
-            else if (schema.anyOf) {
-                schema.anyOf.unshift(internals.nullSchema());
-            }
-            else {
-                schema = {
-                    anyOf: [
-                        internals.nullSchema(),
-                        schema
-                    ]
-                };
-            }
-        }
-
-        // Handle conditionals (whens) by generating multiple possible schemas combined with 'anyOf'
-
-        if (this.$_terms.whens) {
-
-            const base = this.clone();
-            base.$_terms.whens = null;
-
-            const matches = [];
-            for (const when of this.$_terms.whens) {
-                const tests = when.is ? [when] : when.switch;
-                for (let i = 0; i < tests.length; ++i) {
-                    const test = tests[i];
-                    if (test.then) {
-                        matches.push(base.concat(test.then).$_jsonSchema(mode, subOptions));
-                    }
-
-                    if (test.otherwise) {
-                        matches.push(base.concat(test.otherwise).$_jsonSchema(mode, subOptions));
-                    }
-
-                    if (!test.then || (i === tests.length - 1 && !test.otherwise)) {
-                        matches.push(base.$_jsonSchema(mode, subOptions));
-                    }
-                }
-            }
-
-            const results = [];
-            for (const match of matches) {
-                if (!results.some((r) => deepEqual(r, match))) {
-                    results.push(match);
-                }
-            }
-
-            return { anyOf: results };
-        }
-
-        if (isOnly &&
-            rawValids &&
-            rawValids.some((v) => typeof v !== 'symbol') &&
-            !onlyValues.length) {
-
-            return false;
-        }
-
-        if (isOnly && onlyValues && onlyValues.length) {
-            schema = internals.finalizeOnlySchema(this, schema, onlyValues, mode, subOptions);
-        }
-
-        return schema;
+        return JsonSchema.convert(this, mode, options);
     }
 
     // Rules
@@ -1405,414 +1131,6 @@ internals.Base = class {
             }
         };
     }
-};
-
-
-internals.appendCompositeKeyword = function (schema, keyword, value) {
-
-    if (typeof schema === 'boolean') {
-        return schema ? { [keyword]: value } : false;
-    }
-
-    if (schema.allOf) {
-        if (schema[keyword] !== undefined) {
-            schema.allOf.push({ [keyword]: schema[keyword] });
-            delete schema[keyword];
-        }
-
-        schema.allOf.push({ [keyword]: value });
-        return schema;
-    }
-
-    if (schema[keyword] === undefined) {
-        schema[keyword] = value;
-        return schema;
-    }
-
-    schema.allOf = [
-        { [keyword]: schema[keyword] },
-        { [keyword]: value }
-    ];
-
-    delete schema[keyword];
-
-    return schema;
-};
-
-
-internals.finalizeOnlySchema = function (source, schema, values, mode, options) {
-
-    let base = internals.onlyCanRetainBaseSchema(source, values, options.prefs)
-        ? internals.onlyBaseSchema(source, mode, options)
-        : internals.onlyFallbackAnnotations(schema);
-
-    if (typeof base === 'boolean') {
-        base = {};
-    }
-
-    if (schema.$defs) {
-        base.$defs = schema.$defs;
-    }
-
-    if (values.length === 1 &&
-        values[0] === null) {
-
-        base.type = 'null';
-        delete base.enum;
-        return base;
-    }
-
-    base.enum = values;
-
-    const types = internals.onlyTypes(values);
-    if (types) {
-        base.type = types.length === 1 ? types[0] : types;
-    }
-    else {
-        delete base.type;
-    }
-
-    return base;
-};
-
-
-internals.appendType = function (type, addition) {
-
-    const types = Array.isArray(type) ? type.slice() : [type];
-
-    if (!types.includes(addition)) {
-        types.push(addition);
-    }
-
-    return types.length === 1 ? types[0] : types;
-};
-
-
-internals.onlyCanRetainBaseSchema = function (source, values, prefs) {
-
-    const base = internals.onlyBaseClone(source);
-    const baseSchema = base.$_jsonSchema('input', { prefs, $defs: {} });
-    const baseTypes = internals.schemaTypes(baseSchema);
-    let checked = false;
-
-    for (const value of values) {
-        if (value === null) {
-            continue;
-        }
-
-        const valueType = internals.jsonSchemaValueType(value);
-        if (valueType === null) {
-            return false;
-        }
-
-        if (baseTypes &&
-            !baseTypes.has(valueType)) {
-
-            continue;
-        }
-
-        checked = true;
-
-        if (base.validate(value, prefs).error) {
-            return false;
-        }
-    }
-
-    return checked || !baseTypes;
-};
-
-
-internals.onlyBaseSchema = function (source, mode, options) {
-
-    return internals.onlyBaseClone(source).$_jsonSchema(mode, options);
-};
-
-
-internals.onlyBaseClone = function (source) {
-
-    const base = source.clone();
-    base._valids = null;
-    base._invalids = null;
-    delete base._flags.only;
-    return base;
-};
-
-
-internals.onlyFallbackAnnotations = function (schema) {
-
-    if (typeof schema === 'boolean') {
-        return {};
-    }
-
-    const annotations = {};
-    for (const key of internals.onlyFallbackAnnotationKeywords) {
-        if (schema[key] !== undefined) {
-            annotations[key] = schema[key];
-        }
-    }
-
-    return annotations;
-};
-
-
-internals.onlyTypes = function (values) {
-
-    const types = values.map(internals.jsonSchemaValueType);
-    if (types.includes(null)) {
-        return null;
-    }
-
-    return [...new Set(types)];
-};
-
-
-internals.jsonSchemaValues = function (source, values) {
-
-    return internals.uniqueJsonSchemaValues(values.flatMap((value) => internals.jsonSchemaRepresentations(source, value)));
-};
-
-
-internals.jsonSchemaRepresentations = function (source, value) {
-
-    if (source.type === 'date') {
-        if (!(value instanceof Date)) {
-            return [];
-        }
-
-        return internals.jsonSchemaDateValues(source, value);
-    }
-
-    if (value instanceof Date) {
-        return [];
-    }
-
-    return [value];
-};
-
-
-internals.jsonSchemaDateValues = function (source, value) {
-
-    const format = source._flags.format;
-    if (format === 'javascript') {
-        return [value.getTime()];
-    }
-
-    if (format === 'unix') {
-        return [value.getTime() / 1000];
-    }
-
-    if (format === 'iso') {
-        return [value.toISOString()];
-    }
-
-    return [value.toISOString(), value.getTime()];
-};
-
-
-internals.jsonSchemaValueType = function (value) {
-
-    if (value === null) {
-        return 'null';
-    }
-
-    const type = typeof value;
-    return internals.primitiveTypes.has(type) ? type : null;
-};
-
-
-internals.schemaTypes = function (schema) {
-
-    if (typeof schema === 'boolean' ||
-        schema.type === undefined) {
-
-        return null;
-    }
-
-    const types = Array.isArray(schema.type) ? schema.type : [schema.type];
-    return new Set(types);
-};
-
-
-internals.applyExamples = function (source, schema, examples) {
-
-    if (!examples) {
-
-        return;
-    }
-
-    schema.examples = internals.mergeExamples(schema.examples, internals.jsonSchemaAnnotationValue(source, examples));
-};
-
-internals.applyMetas = function (source, schema, metas = []) {
-
-    if (!metas.length) {
-
-        return;
-    }
-
-    for (const meta of metas) {
-        if (!meta ||
-            typeof meta !== 'object' ||
-            Array.isArray(meta)) {
-
-            continue;
-        }
-
-        for (const [key, value] of Object.entries(meta)) {
-            if (key === 'examples') {
-                const merged = internals.mergeExamples(schema.examples, internals.jsonSchemaAnnotationValue(source, value));
-                if (merged !== undefined) {
-                    schema.examples = merged;
-                }
-
-                continue;
-            }
-
-            if (!internals.metaPassthroughKeywords.has(key) ||
-                value === undefined ||
-                schema[key] !== undefined) {
-
-                continue;
-            }
-
-            schema[key] = internals.jsonSchemaAnnotationValue(source, value);
-        }
-    }
-};
-
-
-internals.mergeExamples = function (existing, next) {
-
-    if (!Array.isArray(next) ||
-        !next.length) {
-
-        return existing;
-    }
-
-    const merged = existing ? [...existing] : [];
-    for (const example of next) {
-        if (!merged.some((item) => deepEqual(item, example))) {
-            merged.push(example);
-        }
-    }
-
-    return merged;
-};
-
-
-internals.uniqueJsonSchemaValues = function (values) {
-
-    return internals.mergeExamples([], values);
-};
-
-
-internals.jsonSchemaAnnotationValue = function (source, value) {
-
-    if (value instanceof Date) {
-        return internals.jsonSchemaAnnotationDateValue(source, value);
-    }
-
-    if (Array.isArray(value)) {
-        return value.map((item) => internals.jsonSchemaAnnotationValue(source, item));
-    }
-
-    if (internals.isPlainObject(value)) {
-        const copy = {};
-        for (const [key, item] of Object.entries(value)) {
-            copy[key] = internals.jsonSchemaAnnotationValue(source, item);
-        }
-
-        return copy;
-    }
-
-    return value;
-};
-
-
-internals.jsonSchemaDefaultValue = function (source, value) {
-
-    if (source?.type === 'date' &&
-        value === 'now') {
-
-        return undefined;
-    }
-
-    return internals.jsonSchemaAnnotationValue(source, value);
-};
-
-
-internals.jsonSchemaAnnotationDateValue = function (source, value) {
-
-    if (source?.type === 'date') {
-        const format = source._flags.format;
-        if (format === 'javascript') {
-            return value.getTime();
-        }
-
-        if (format === 'unix') {
-            return value.getTime() / 1000;
-        }
-    }
-
-    return value.toISOString();
-};
-
-
-internals.isPlainObject = function (value) {
-
-    if (!value ||
-        typeof value !== 'object') {
-
-        return false;
-    }
-
-    const prototype = Object.getPrototypeOf(value);
-    return prototype === Object.prototype || prototype === null;
-};
-
-
-internals.schemaCanMatchNull = function (schema) {
-
-    if (schema === false) {
-        return false;
-    }
-
-    if (schema === true) {
-        return true;
-    }
-
-    if (schema.enum) {
-        return schema.enum.includes(null);
-    }
-
-    if (schema.type !== undefined) {
-        const types = Array.isArray(schema.type) ? schema.type : [schema.type];
-        return types.includes('null');
-    }
-
-    if (schema.anyOf) {
-        return schema.anyOf.some(internals.schemaCanMatchNull);
-    }
-
-    if (schema.oneOf) {
-        return schema.oneOf.some(internals.schemaCanMatchNull);
-    }
-
-    if (schema.allOf) {
-        return schema.allOf.every(internals.schemaCanMatchNull);
-    }
-
-    return true;
-};
-
-
-internals.jsonSchemaType = function (schema) {
-
-    if (internals.standardTypes.has(schema.type)) {
-        return schema.type;
-    }
-
-    return schema._definition?.jsonSchemaType || null;
 };
