+++ lib/types/alternatives.js
@@ -6,6 +6,7 @@ const Any = require('./any');
 const Common = require('../common');
 const Compile = require('../compile');
 const Errors = require('../errors');
+const JsonAlternatives = require('../json-schema/alternatives');
 const Ref = require('../ref');
 
 
@@ -148,29 +149,7 @@ module.exports = Any.extend({
 
     jsonSchema(schema, res, mode, options) {
 
-        const matches = [];
-
-        // Collect all alternative schemas from 'matches' term
-
-        for (const match of schema.$_terms.matches) {
-            if (match.schema) {
-                matches.push(match.schema.$_jsonSchema(mode, options));
-            }
-            else {
-                // Handle conditional matches (when/switch)
-
-                const tests = match.is ? [match] : match.switch;
-                for (const test of tests) {
-                    if (test.then) {
-                        matches.push(test.then.$_jsonSchema(mode, options));
-                    }
-
-                    if (test.otherwise) {
-                        matches.push(test.otherwise.$_jsonSchema(mode, options));
-                    }
-                }
-            }
-        }
+        const matches = JsonAlternatives.matches(schema, mode, options);
 
         if (matches.length) {
             delete res.type;
