+++ lib/types/binary.js
@@ -4,6 +4,7 @@ const { assert } = require('@hapi/hoek');
 
 const Any = require('./any');
 const Common = require('../common');
+const JsonBinary = require('../json-schema/binary');
 
 
 const internals = {};
@@ -37,7 +38,7 @@ module.exports = Any.extend({
 
         res.type = 'string';
 
-        const contentEncoding = internals.contentEncoding(schema._flags.encoding);
+        const contentEncoding = JsonBinary.contentEncoding(schema._flags.encoding);
         if (contentEncoding) {
             res.contentEncoding = contentEncoding;
         }
@@ -126,27 +127,3 @@ module.exports = Any.extend({
         'binary.min': '{{#label}} must be at least {{#limit}} bytes'
     }
 });
-
-
-internals.contentEncoding = function (encoding) {
-
-    // JSON Schema's contentEncoding annotation follows RFC 4648 / MIME transfer
-    // encoding names, not Node's full Buffer encoding namespace. Map only the
-    // binary transfer encodings Joi can express here and omit charset-style
-    // encodings such as utf8/latin1 that have no honest contentEncoding value.
-    if (!encoding) {
-        return undefined;
-    }
-
-    if (encoding === 'hex') {
-        return 'base16';
-    }
-
-    if (encoding === 'base64' ||
-        encoding === 'base64url') {
-
-        return encoding;
-    }
-
-    return undefined;
-};
